Change Xid_New_Performance test timeout to 2 seconds (debug and release).
The original code no longer runs in the timeout allowed, even though it used to. This probably indicates it was always 'too tight' and would fail on different machines. The recent code changes shouldn't have much impact on performance since they're in a method that only runs once (value is cached in a static variable) and if anything should actually be marginally faster. Whatever regression has happened doesn't appear to be strictly code related, so increasing the timeout seems reasonable.